requirements.txt
cassandra-driver==3.29.2
click==8.1.7
cqlsh==6.2.0
geomet==0.2.1.post1
kafka-python==2.0.2
pure-sasl==0.6.2
six==1.16.0
wcwidth==0.2.13
